+++ toptalk_user/src/main/java/com/toptalk/user/controller/UserController.java
@@ -5,10 +5,14 @@ import com.toptalk.user.service.UserService;
 import entity.PageResult;
 import entity.Result;
 import entity.StatusCode;
+import io.jsonwebtoken.Claims;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.web.bind.annotation.*;
+import util.JwtUtil;
 
+import javax.servlet.http.HttpServletRequest;
+import java.util.HashMap;
 import java.util.Map;
 /**
  * 控制器层
@@ -22,6 +26,34 @@ public class UserController {
 
 	@Autowired
 	private UserService userService;
+
+	@Autowired
+    private HttpServletRequest request;
+
+	@Autowired
+	private JwtUtil jwtUtil;
+
+	/**
+	 * 用户根据手机号登陆;
+	 * @param mobile
+	 * @param password
+	 * @return
+	 */
+	@RequestMapping(value = "/login",method = RequestMethod.POST)
+	public Result login(String mobile, String password){
+		User user = userService.findByMobileAndPassword(mobile,password);
+		if(null != user){
+			//创建token信息,并返回给前端;;
+			String token = jwtUtil.createJWT(user.getId(),user.getNickname(),"user");
+			Map map = new HashMap();
+			map.put("token",token);
+			map.put("name",user.getNickname());
+			map.put("avatar",user.getAvatar());
+			return new Result(true,StatusCode.OK,"登录成功!",map);
+		}else {
+			return new Result(false,StatusCode.LOGINERROR,"用户名或密码错误!");
+		}
+	}
 
 	/**
 	 * 用户注册;
@@ -111,11 +143,16 @@ public class UserController {
 	}
 	
 	/**
-	 * 删除
+	 * 删除,需要验证是否有管理员权限;
 	 * @param id
 	 */
 	@RequestMapping(value="/{id}",method= RequestMethod.DELETE)
 	public Result delete(@PathVariable String id ){
+		//判断是否有管理员权限;
+		Claims claims = (Claims)request.getAttribute("admin_claims");
+		if (null == claims){
+			return new Result(false,StatusCode.ACCESSERROR,"操作权限不足!");
+		}
 		userService.deleteById(id);
 		return new Result(true,StatusCode.OK,"删除成功");
 	}
